@@ -7,8 +7,11 @@
 //
 
 #import "AppDelegate.h"
+#import "FLNativeView.h"
 #import "UIViewControllerDemo.h"
-#import "DemoRouter.h"
+#import "NativeViewController.h"
+#import "MyFlutterBoostDelegate.h"
+
 #import <flutter_boost/FlutterBoost.h>
 
 @interface AppDelegate ()
@@ -19,51 +22,59 @@
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
 {
-    DemoRouter *router = [DemoRouter sharedRouter];
-    [FlutterBoostPlugin.sharedInstance startFlutterWithPlatform:router
-                                                        onStart:^(FlutterEngine *fvc) {
-                                                            
-                                                        }];
-    
     self.window = [[UIWindow alloc] initWithFrame: [UIScreen mainScreen].bounds];
-    
-    
     [self.window makeKeyAndVisible];
-    
-   
+    //默认方法
+    MyFlutterBoostDelegate* delegate=[[MyFlutterBoostDelegate alloc ] init];
+    [[FlutterBoost instance] setup:application delegate:delegate callback:^(FlutterEngine *engine) {
+        NSObject<FlutterPluginRegistrar>* registrar = [engine registrarForPlugin:@"plugin-name"];
+        FLNativeViewFactory* factory = [[FLNativeViewFactory alloc] initWithMessenger:registrar.messenger];
+        [[engine registrarForPlugin:@"<plugin-name>"] registerViewFactory:factory withId:@"<simple-text-view>"];
+    } ];
+
+    //下面是自定义option参数的方法
+
+//    FlutterBoostSetupOptions* options = [FlutterBoostSetupOptions createDefault];
+//    options.dartEntryPoint = @"main2";
+//
+//    [[FlutterBoost instance] setup:application delegate:delegate callback:^(FlutterEngine *engine) {
+//    } options:options];
+
     UIViewControllerDemo *vc = [[UIViewControllerDemo alloc] initWithNibName:@"UIViewControllerDemo" bundle:[NSBundle mainBundle]];
     vc.tabBarItem = [[UITabBarItem alloc] initWithTitle:@"hybrid" image:nil tag:0];
-   
-    
-    FLBFlutterViewContainer *fvc = FLBFlutterViewContainer.new;
-    [fvc setName:@"tab" params:@{}];
+
+    FBFlutterViewContainer *fvc = FBFlutterViewContainer.new ;
+
+    [fvc setName:@"tab_friend" uniqueId:nil params:@{} opaque:YES];
     fvc.tabBarItem = [[UITabBarItem alloc] initWithTitle:@"flutter_tab" image:nil tag:1];
-    
-    
+
+
     UITabBarController *tabVC = [[UITabBarController alloc] init];
-    UINavigationController *rvc = [[UINavigationController alloc] initWithRootViewController:tabVC];
-    
-   
-    router.navigationController = rvc;
-    
     tabVC.viewControllers = @[vc,fvc];
-    
+
+
+    UINavigationController *rvc = [[UINavigationController alloc] initWithRootViewController:tabVC];
+
+    delegate.navigationController=rvc;
+
+
     self.window.rootViewController = rvc;
-    
-    
-  
-    
-    
-    
-    
+
     UIButton *nativeButton = [UIButton buttonWithType:UIButtonTypeCustom];
-    nativeButton.frame = CGRectMake(self.window.frame.size.width * 0.5 - 50, 200, 100, 45);
+    nativeButton.frame = CGRectMake(self.window.frame.size.width * 0.5 - 50, 200, 100, 40);
     nativeButton.backgroundColor = [UIColor redColor];
     [nativeButton setTitle:@"push native" forState:UIControlStateNormal];
     [nativeButton addTarget:self action:@selector(pushNative) forControlEvents:UIControlEventTouchUpInside];
     [self.window addSubview:nativeButton];
-    
-    
+
+    UIButton *pushEmbeded = [UIButton buttonWithType:UIButtonTypeCustom];
+    pushEmbeded.frame = CGRectMake(self.window.frame.size.width * 0.5 - 70, 150, 140, 40);
+    pushEmbeded.backgroundColor = [UIColor redColor];
+    [pushEmbeded setTitle:@"push embedded" forState:UIControlStateNormal];
+    [pushEmbeded addTarget:self action:@selector(pushEmbeded) forControlEvents:UIControlEventTouchUpInside];
+    [self.window addSubview:pushEmbeded];
+
+
     return YES;
 }
 
@@ -73,7 +84,13 @@
     UIViewControllerDemo *vc = [[UIViewControllerDemo alloc] initWithNibName:@"UIViewControllerDemo" bundle:[NSBundle mainBundle]];
     [nvc pushViewController:vc animated:YES];
 }
-
+//
+- (void)pushEmbeded
+{
+    UINavigationController *nvc = (id)self.window.rootViewController;
+    UIViewController *vc = [[NativeViewController alloc] init];
+    [nvc pushViewController:vc animated:YES];
+}
 
 - (void)applicationWillResignActive:(UIApplication *)application {
     // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
